Propagate ONNX #8149 internal version type widening (int64_t plumbing) (#3147)
* Initial plan

* propagate onnx/onnx#8149 internal version type widening


---------

+++ onnx_light/onnx_lib/shape_inference/implementation.h
@@ -106,7 +106,8 @@ struct GraphInferenceContext {
       std::unordered_map<std::string, int> opset_imports_in, SymbolTable *symbol_table_in = nullptr,
       const ModelLocalFunctionsMap &model_local_functions_in = {},
       const ISchemaRegistry *schema_registry_in = OpSchemaRegistry::Instance(),
-      DataValueMap *generated_shape_data_by_name_in = nullptr, const int ir_version_in = IR_VERSION)
+      DataValueMap *generated_shape_data_by_name_in = nullptr,
+      const int64_t ir_version_in = IR_VERSION)
       : outer_scope_value_types_by_name{&outer_scope_value_types_by_name_in},
         opset_imports{std::move(opset_imports_in)}, symbol_table{symbol_table_in},
         model_local_functions{model_local_functions_in}, schema_registry{schema_registry_in},
@@ -118,7 +119,7 @@ struct GraphInferenceContext {
   const ModelLocalFunctionsMap &model_local_functions;
   const ISchemaRegistry *schema_registry;
   DataValueMap *generated_shape_data_by_name;
-  const int ir_version;
+  const int64_t ir_version;
 };
 
 class GraphInferencerImpl : public GraphInferencer {
